backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
psycopg = {extras = ["binary"], version = "^3.2.3"}
newsapi-python = "^0.2.7"
openai = "^1.59.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
reportlab = "^4.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
